+++ src/main/java/phanes/replay/user/strategy/impl/GoogleLoginStrategy.java
@@ -1,4 +1,4 @@
-package phanes.replay.user.strategy;
+package phanes.replay.user.strategy.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -12,6 +12,7 @@ import phanes.replay.user.domain.enums.SocialType;
 import phanes.replay.user.dto.oauth.GoogleProfileResponse;
 import phanes.replay.user.dto.oauth.GoogleTokenResponse;
 import phanes.replay.user.dto.oauth.SocialProfile;
+import phanes.replay.user.strategy.SocialLoginStrategy;
 
 @Component
 @RequiredArgsConstructor
