+++ src/main/java/phanes/replay/user/strategy/impl/NaverLoginStrategy.java
@@ -1,4 +1,4 @@
-package phanes.replay.user.strategy;
+package phanes.replay.user.strategy.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -11,6 +11,7 @@ import phanes.replay.user.dto.oauth.NaverProfile;
 import phanes.replay.user.dto.oauth.NaverProfileResponse;
 import phanes.replay.user.dto.oauth.NaverTokenResponse;
 import phanes.replay.user.dto.oauth.SocialProfile;
+import phanes.replay.user.strategy.SocialLoginStrategy;
 
 @Component
 @RequiredArgsConstructor
